+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:targetApi="31">
-        <activity
-            android:name=".activities.ExportActivity"
-            android:exported="false" />
+
         <activity
             android:name=".activities.ReportActivity"
             android:exported="false" />
