--- app/src/main/res/layout/activity_export.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".activities.ExportActivity">
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_export_all"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Save All"
-        />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_export_range"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Select Range"
-        />
-</androidx.appcompat.widget.LinearLayoutCompat>
